// packages/core/src/generated/linear-easings.ts
export const springCalculations: Record<SpringCurveKey, Value> = {
  HEAVY: {
    easingValue:
      'linear(0, 0.013 0.6%, 0.05 1.2%, 0.2 2.5%, 0.949 6.7%, 1.2 8.4%, 1.286 9.2%, 1.35 10%, 1.392 10.8%, 1.411 11.6%, 1.411 12.2%, 1.401 12.8%, 1.343 14.2%, 1.258 15.5%, 1.016 18.7%, 0.914 20.4%, 0.856 21.9%, 0.831 23.5%, 0.834 24.7%, 0.858 26.1%, 0.996 30.7%, 1.037 32.4%, 1.06 33.9%, 1.07 35.4%, 1.061 37.7%, 0.989 43.8%, 0.971 47.2%, 1.012 59.1%, 0.995 70.8%, 1)',
    sampledPoints: [
      {
        x: 0,
        y: 100,
      },
      {
        x: 0.6,
        y: 98.7,
      },
      {
        x: 1.2,
        y: 95,
      },
      {
        x: 2.5,
        y: 80,
      },
      {
        x: 6.7,
        y: 5.1,
      },
      {
        x: 8.4,
        y: -20,
      },
      {
        x: 9.2,
        y: -28.6,
      },
      {
        x: 10,
        y: -35,
      },
      {
        x: 10.8,
        y: -39.2,
      },
      {
        x: 11.6,
        y: -41.1,
      },
      {
        x: 12.2,
        y: -41.1,
      },
      {
        x: 12.8,
        y: -40.1,
      },
      {
        x: 14.2,
        y: -34.3,
      },
      {
        x: 15.5,
        y: -25.8,
      },
      {
        x: 18.7,
        y: -1.6,
      },
      {
        x: 20.4,
        y: 8.6,
      },
      {
        x: 21.9,
        y: 14.4,
      },
      {
        x: 23.5,
        y: 16.9,
      },
      {
        x: 24.7,
        y: 16.6,
      },
      {
        x: 26.1,
        y: 14.2,
      },
      {
        x: 30.7,
        y: 0.4,
      },
      {
        x: 32.4,
        y: -3.7,
      },
      {
        x: 33.9,
        y: -6,
      },
      {
        x: 35.4,
        y: -7,
      },
      {
        x: 37.7,
        y: -6.1,
      },
      {
        x: 43.8,
        y: 1.1,
      },
      {
        x: 47.2,
        y: 2.9,
      },
      {
        x: 59.1,
        y: -1.2,
      },
      {
        x: 70.8,
        y: 0.5,
      },
      {
        x: 100,
        y: 0,
      },
    ],
  },
// ...
};
export const bounceCalculations: Record<BounceCurveKey, Value> = {
  FIRM: {
    easingValue:
      'linear(0, 0.223 11.7%, 0.392 18.4%, 0.619 24.8%, 0.999 33.3%, 0.748 40%, 0.691 42.7%, 0.672 45.3%, 0.69 47.8%, 0.743 50.4%, 0.999 57.7%, 0.883 61.8%, 0.856 63.6%, 0.848 65.3%, 0.855 67%, 0.879 68.8%, 0.999 74.5%, 0.953 77.5%, 0.94 80.2%, 0.95 82.7%, 1 88.2%, 0.987 91.9%, 1)',
    sampledPoints: [
      {
        x: 0,
        y: 100,
      },
      {
        x: 11.7,
        y: 77.7,
      },
      {
        x: 18.4,
        y: 60.8,
      },
      {
        x: 24.8,
        y: 38.1,
      },
      {
        x: 33.3,
        y: 0.1,
      },
      {
        x: 40,
        y: 25.2,
      },
      {
        x: 42.7,
        y: 30.9,
      },
      {
        x: 45.3,
        y: 32.8,
      },
      {
        x: 47.8,
        y: 31,
      },
      {
        x: 50.4,
        y: 25.7,
      },
      {
        x: 57.7,
        y: 0.1,
      },
      {
        x: 61.8,
        y: 11.7,
      },
      {
        x: 63.6,
        y: 14.4,
      },
      {
        x: 65.3,
        y: 15.2,
      },
      {
        x: 67,
        y: 14.5,
      },
      {
        x: 68.8,
        y: 12.1,
      },
      {
        x: 74.5,
        y: 0.1,
      },
      {
        x: 77.5,
        y: 4.7,
      },
      {
        x: 80.2,
        y: 6,
      },
      {
        x: 82.7,
        y: 5,
      },
      {
        x: 88.2,
        y: 0,
      },
      {
        x: 91.9,
        y: 1.3,
      },
      {
        x: 100,
        y: 0,
      },
    ],
  },
  SOFT: {
    easingValue: 'linear(0, 1 44.7%, 0.898 51.8%, 0.874 55.1%, 0.866 58.4%, 0.888 64.3%, 1 77.4%, 0.98 84.5%, 1)',
    sampledPoints: [
      {
        x: 0,
        y: 100,
      },
      {
        x: 44.7,
        y: 0,
      },
      {
        x: 51.8,
        y: 10.2,
      },
      {
        x: 55.1,
        y: 12.6,
      },
      {
        x: 58.4,
        y: 13.4,
      },
      {
        x: 64.3,
        y: 11.2,
      },
      {
        x: 77.4,
        y: 0,
      },
      {
        x: 84.5,
        y: 2,
      },
      {
        x: 100,
        y: 0,
      },
    ],
  },
// ...
  SUBTLE: {
    easingValue:
      'linear(0, 0.417 25.5%, 0.867 49.4%, 1 57.7%, 0.925 65.1%, 0.908 68.6%, 0.902 72.2%, 0.916 78.2%, 0.988 92.1%, 1)',
    sampledPoints: [
      {
        x: 0,
        y: 100,
      },
      {
        x: 25.5,
        y: 58.3,
      },
      {
        x: 49.4,
        y: 13.3,
      },
      {
        x: 57.7,
        y: 0,
      },
      {
        x: 65.1,
        y: 7.5,
      },
      {
        x: 68.6,
        y: 9.2,
      },
      {
        x: 72.2,
        y: 9.8,
      },
      {
        x: 78.2,
        y: 8.4,
      },
      {
        x: 92.1,
        y: 1.2,
      },
      {
        x: 100,
        y: 0,
      },
    ],
  },
// ...
};
export const wiggleCalculations: Record<WiggleCurveKey, Value> = {
  SUBTLE: {
    easingValue:
      'linear(0, 0.409 5.7%, 0.68 10.7%, 0.838 15.5%, 0.877 17.9%, 0.89 20.3%, 0.883 22.1%, 0.862 24%, 0.775 28.1%, 0.103 46.8%, -0.03 51.9%, -0.11 56.9%, -0.142 61.2%, -0.143 65.9%, -0.124 70.2%, -0.021 86%, -0.004 92.1%, 0)',
    sampledPoints: [
      {
        x: 0,
        y: 100,
      },
      {
        x: 5.7,
        y: 59.1,
      },
      {
        x: 10.7,
        y: 32,
      },
      {
        x: 15.5,
        y: 16.2,
      },
      {
        x: 17.9,
        y: 12.3,
      },
      {
        x: 20.3,
        y: 11,
      },
      {
        x: 22.1,
        y: 11.7,
      },
      {
        x: 24,
        y: 13.8,
      },
      {
        x: 28.1,
        y: 22.5,
      },
      {
        x: 46.8,
        y: 89.7,
      },
      {
        x: 51.9,
        y: 103,
      },
      {
        x: 56.9,
        y: 111,
      },
      {
        x: 61.2,
        y: 114.2,
      },
      {
        x: 65.9,
        y: 114.3,
      },
      {
        x: 70.2,
        y: 112.4,
      },
      {
        x: 86,
        y: 102.1,
      },
      {
        x: 92.1,
        y: 100.4,
      },
      {
        x: 100,
        y: 100,
      },
    ],
  },
// ...
};
export const overshootCalculations: Record<OvershootStyleKey, Record<OvershootCurveKey, Value>> = {
// ...
  OUT: {
    SOFT: {
      easingValue:
        'linear(0, 0.402 7.4%, 0.711 15.3%, 0.929 23.7%, 1.008 28.2%, 1.067 33%, 1.099 36.9%, 1.12 41%, 1.13 45.4%, 1.13 50.1%, 1.111 58.5%, 1.019 83.2%, 1.004 91.3%, 1)',
      sampledPoints: [
        {
          x: 0,
          y: 100,
        },
        {
          x: 7.4,
          y: 59.8,
        },
        {
          x: 15.3,
          y: 28.9,
        },
        {
          x: 23.7,
          y: 7.1,
        },
        {
          x: 28.2,
          y: -0.8,
        },
        {
          x: 33,
          y: -6.7,
        },
        {
          x: 36.9,
          y: -9.9,
        },
        {
          x: 41,
          y: -12,
        },
        {
          x: 45.4,
          y: -13,
        },
        {
          x: 50.1,
          y: -13,
        },
        {
          x: 58.5,
          y: -11.1,
        },
        {
          x: 83.2,
          y: -1.9,
        },
        {
          x: 91.3,
          y: -0.4,
        },
        {
          x: 100,
          y: 0,
        },
      ],
    },
// ...
  },
  IN_OUT: {
// ...
    FIRM: {
      easingValue:
        'linear(0, -0.008 4.1%, -0.035 8.3%, -0.179 21.1%, -0.216 25.3%, -0.228 29.5%, -0.208 33.3%, -0.121 38.2%, 0.036 42.7%, 0.261 46.8%, 0.881 55.6%, 1.058 59.7%, 1.172 64.1%, 1.222 68.6%, 1.228 71.1%, 1.221 73.7%, 1.183 78.5%, 1.034 91.8%, 1.008 95.9%, 1)',
      sampledPoints: [
        {
          x: 0,
          y: 100,
        },
        {
          x: 4.1,
          y: 100.8,
        },
        {
          x: 8.3,
          y: 103.5,
        },
        {
          x: 21.1,
          y: 117.9,
        },
        {
          x: 25.3,
          y: 121.6,
        },
        {
          x: 29.5,
          y: 122.8,
        },
        {
          x: 33.3,
          y: 120.8,
        },
        {
          x: 38.2,
          y: 112.1,
        },
        {
          x: 42.7,
          y: 96.4,
        },
        {
          x: 46.8,
          y: 73.9,
        },
        {
          x: 55.6,
          y: 11.9,
        },
        {
          x: 59.7,
          y: -5.8,
        },
        {
          x: 64.1,
          y: -17.2,
        },
        {
          x: 68.6,
          y: -22.2,
        },
        {
          x: 71.1,
          y: -22.8,
        },
        {
          x: 73.7,
          y: -22.1,
        },
        {
          x: 78.5,
          y: -18.3,
        },
        {
          x: 91.8,
          y: -3.4,
        },
        {
          x: 95.9,
          y: -0.8,
        },
        {
          x: 100,
          y: 0,
        },
      ],
    },
// ...
  },
};